add name selectors to trace and debug for logp
